+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/application/IFizzBuzzCommand.java
@@ -1,4 +1,4 @@
-package fizzbuzz;
+package fizzbuzz.application;
 
 public interface IFizzBuzzCommand {
     void execute(int arg);
