+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/domain/model/FizzBuzzValues.java
@@ -1,4 +1,4 @@
-package fizzbuzz;
+package fizzbuzz.domain.model;
 
 import java.util.ArrayList;
 import java.util.List;
