+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/domain/type/FizzBuzzType02.java
@@ -1,4 +1,6 @@
-package fizzbuzz;
+package fizzbuzz.domain.type;
+
+import fizzbuzz.domain.model.FizzBuzzValue;
 
 public class FizzBuzzType02 extends IFizzBuzzType {
     @Override
