+++ dev/20190402/k2works/sam-app/HelloWorldFunction/src/main/java/fizzbuzz/domain/type/FizzBuzzType03.java
@@ -1,4 +1,6 @@
-package fizzbuzz;
+package fizzbuzz.domain.type;
+
+import fizzbuzz.domain.model.FizzBuzzValue;
 
 public class FizzBuzzType03 extends IFizzBuzzType {
     @Override
